+++ stub/ray.php
@@ -41,6 +41,11 @@ return [
     'send_duplicate_queries_to_ray' => env('SEND_DUPLICATE_QUERIES_TO_RAY', false),
 
     /*
+     * When enabled, slow queries will automatically be sent to Ray.
+     */
+    'send_slow_queries_to_ray' => env('SEND_SLOW_QUERIES_TO_RAY', false),
+    
+    /*
     * When enabled, all requests made to this app will automatically be sent to Ray.
     */
     'send_requests_to_ray' => env('SEND_REQUESTS_TO_RAY', false),